Part of #99. Various connection handling improvements
Various connection handling improvements:-
- Correctly clean up connections and don't leave hanging references within zephyr
- Logging all internal connection state
- Improved logging for multiple parameters, and tests pass
- Successfully call application's didMeasure withPayload after initial payload read
- BUG: No subsequent didMeasure withPayload calls (likely due to duplicate Devices in BLE DB)

+++ herald/include/herald/data/sensor_logger.h
@@ -177,10 +177,12 @@ namespace {
     os << std::int16_t(value);
   }
   
+  /// \brief Prints an int value to the stream
+  /// \note Also covers int32_t and int64_t
   [[maybe_unused]]
   void tprintValue(std::stringstream& os, int value)
   {
-    os << std::int16_t(value);
+    os << std::int64_t(value);
   }
   
   [[maybe_unused]]
